SIMPLE ORDERING SYSTEM IDEA
This includes the SimpleOrderingSystem.py file.

To RUN the code - Simply Run it as normal.

What it does:
User will be shown a menu of grocery items with their respective prices and stock. The user can then choose the grocery item of their liking and set the amount to whatever they like, as long as the amount is not greater than the stock provided. The program ends when the user quits the program.