Added StockTracking.py
How to run file - Run as normal

How it works: 
The user is greeted with a prompt that says "Stock Tracking" and is shown a list of items with their prices and stock available.

The user can navigate to see the stock of each individual item. It should be able to display the list of items either in: Alphabetical Order or the amount of stock available (Either in Ascending or Descending form). There should also be a prompt like that for the user to choose via numbers that corresponds to the chosen arrangement for the item list. 

Step 3: After the user is done in checking the stock, the user can exit the program.